OpenID4VP: Move out ClientIdScheme into separate file
