+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpCombinedProtocolTwoStepTest.kt
@@ -31,7 +31,7 @@ class OpenId4VpCombinedProtocolTwoStepTest : FreeSpec({
         )
         verifierOid4vp = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
-            clientIdScheme = OpenId4VpVerifier.ClientIdScheme.RedirectUri(clientId),
+            clientIdScheme = ClientIdScheme.RedirectUri(clientId),
         )
     }
